Enhance HnStoryContentCard by integrating comment and scroll stores for improved text collapse functionality. Update prop structure to include firstCommentId, simplifying the interaction logic. Remove unused imports and streamline the component's state management.
