+++ src/features/comments/HnStoryContentCard.tsx
@@ -1,7 +1,11 @@
 import { Show } from "solid-js";
 
 import { ArrowUpRightFromSquare } from "~/components/Icon";
-import { useColorMapStore } from "~/contexts/AppDataContext";
+import {
+  useColorMapStore,
+  useCommentStore,
+  useScrollStore,
+} from "~/contexts/AppDataContext";
 import { processHtmlAndTruncateAnchorText } from "~/lib/processHtmlAndTruncateAnchorText";
 import { cn, getDomain, shareSafely, timeSince } from "~/lib/utils";
 
@@ -9,22 +13,48 @@ import type { HnItem } from "~/models/interfaces";
 
 type HnStoryContentCardProps = {
   story?: HnItem | undefined;
-  isTextOpen: boolean;
-  onToggleText: () => void;
+  firstCommentId?: number;
 };
 
 export const HnStoryContentCard = (props: HnStoryContentCardProps) => {
   const colorMapStore = useColorMapStore();
+  const commentStore = useCommentStore();
+  const scrollStore = useScrollStore();
 
   const author = () => props.story?.by;
   const score = () => props.story?.score;
   const time = () => props.story?.time;
   const url = () => props.story?.url;
   const hasText = () => props.story?.text !== undefined;
+  const storyId = () => props.story?.id;
   const textHtml = () =>
     processHtmlAndTruncateAnchorText(props.story?.text || "");
   const flashColor = () =>
     colorMapStore.colorMap()[author() ?? ""] ?? "hsl(30, 80%, 65%)";
+
+  const isTextOpen = () => {
+    if (storyId() === undefined) {
+      return true;
+    }
+    return commentStore.collapsedTimestamps[storyId()!] === undefined;
+  };
+
+  function handleStoryTextClick() {
+    if (!props.story?.text || storyId() === undefined) {
+      return;
+    }
+
+    const newIsCollapsed = !!isTextOpen();
+    commentStore.updateCollapsedState(storyId()!, newIsCollapsed);
+
+    // scroll to first comment if it exists
+    // schedule out 200ms to allow the collapse animation to finish
+    setTimeout(() => {
+      if (newIsCollapsed && props.firstCommentId) {
+        scrollStore.setScrollToId(props.firstCommentId);
+      }
+    }, 100);
+  }
 
   const handleShareClick = async (event: MouseEvent) => {
     event.preventDefault();
@@ -38,11 +68,11 @@ export const HnStoryContentCard = (props: HnStoryContentCardProps) => {
       class={cn(
         {
           "rounded-tl pr-2 pl-4": hasText(),
-          collapsed: !props.isTextOpen,
+          collapsed: !isTextOpen(),
         },
         "bp3-card"
       )}
-      onClick={() => props.onToggleText()}
+      onClick={handleStoryTextClick}
       style={{
         "--flash-color": flashColor(),
         "padding-left": "16px",
@@ -80,7 +110,7 @@ export const HnStoryContentCard = (props: HnStoryContentCardProps) => {
         </button>
       </div>
 
-      <Show when={hasText() && props.isTextOpen}>
+      <Show when={hasText() && isTextOpen()}>
         <div>
           {/*  eslint-disable-next-line solid/no-innerhtml */}
           <p class="user-text break-words" innerHTML={textHtml()} />
